Revert "Move authentication system to Publish::Authentication namespace"
This reverts commit 7e89fc8586ef2541ea1fa3746d8d7d7a78851e78.
