--- app/controllers/publish/authentication/sign_in_controller.rb
@@ -1,18 +1,0 @@
-# frozen_string_literal: true
-
-module Publish
-  module Authentication
-    class SignInController < ApplicationController
-      skip_before_action :authenticate
-      skip_after_action :verify_authorized
-
-      def index
-        if AuthenticationService.magic_link?
-          redirect_to magic_links_path
-        else
-          render "index"
-        end
-      end
-    end
-  end
-end
