+++ app/controllers/sign_in_controller.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class SignInController < ApplicationController
+  skip_before_action :authenticate
+
+  def index
+    if AuthenticationService.magic_link?
+      redirect_to magic_links_path
+    else
+      render "index"
+    end
+  end
+end
