+++ app/controllers/support/application_controller.rb
@@ -2,7 +2,6 @@
 
 module Support
   class ApplicationController < ::ApplicationController
-    include Publish::Authentication
     layout "support"
     before_action :check_user_is_admin
