+++ app/lib/blazer_admin_constraint.rb
@@ -2,7 +2,7 @@
 
 class BlazerAdminConstraint
   def matches?(request)
-    signin_user = Publish::Authentication::UserSession.load_from_session(request.session)
+    signin_user = UserSession.load_from_session(request.session)
     return false if signin_user.blank?
 
     User.with_blazer_access.kept.exists?(email: signin_user.email)
