+++ app/lib/system_admin_constraint.rb
@@ -6,7 +6,7 @@ class SystemAdminConstraint
   end
 
   def system_admin?(request)
-    signin_user = Publish::Authentication::UserSession.load_from_session(request.session)
+    signin_user = UserSession.load_from_session(request.session)
     signin_user.present? && User.admins.kept.exists?(email: signin_user.email)
   end
 end
