+++ app/views/magic_links/magic_link_sent.html.erb
