+++ app/views/magic_links/new.html.erb
