+++ app/views/personas/index.html.erb
