+++ config/locales/en.yml
@@ -256,6 +256,10 @@ en:
       check: "Check your answers"
       delete: "Are you sure you want to remove this user?"
       edit: "Personal details"
+  publish_authentication:
+    magic_link:
+      invalid_token: "Magic link could not be verified, please request a new one"
+      expired: "Magic link has expired, please request a new one"
   navigation_bar:
     courses: "Courses"
     schools: "Schools"
@@ -937,11 +941,6 @@ en:
               invalid: "Enter a real postcode"
             urn:
               format: "URN must be 5 or 6 numbers"
-        publish/authentication/magic_link_form:
-          attributes:
-            email:
-              blank: "Enter an email adddress"
-              invalid: "Enter a valid email address"
         publish/school_form:
           attributes:
             location_name:
@@ -1025,6 +1024,11 @@ en:
           attributes:
             terms_accepted:
               accepted: "You must accept the terms and conditions to continue"
+        publish/authentication/magic_link_form:
+          attributes:
+            email:
+              blank: "Enter an email adddress"
+              invalid: "Enter a valid email address"
         publish/providers/schools/search_form:
           attributes:
             query:
