+++ config/locales/en/publish.yml
@@ -1,9 +1,5 @@
 en:
   publish:
-    authentication:
-      magic_link:
-        invalid_token: "Magic link could not be verified, please request a new one"
-        expired: "Magic link has expired, please request a new one"
     courses:
       index:
         ratifying_provider: Accredited provider
