+++ spec/system/auth/provider_user_signs_in_spec.rb
