+++ spec/system/auth/support_user_signs_in_spec.rb
